refactor(graph-canvas): move rendering modules to representation
Fixes #466
