+++ src/engine/index.ts
@@ -1,6 +1,4 @@
 export { buildGraph, getNodeDegrees, getEdgeDescription } from './graph';
-export { createGraphNetwork, computeGraphPositions, buildVisNode } from './createGraphNetwork';
-export { ICON_NODE_SHAPE } from './nodeRenderer';
 export {
   loadAuthoredContent,
   loadRepoContent,
