+++ src/representation/graph-canvas/__tests__/nodeRenderer-theme.test.ts
