+++ src/representation/graph-canvas/__tests__/nodeRenderer.test.ts
