+++ src/representation/graph-canvas/__tests__/reports-to-layout.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
-import type { KBGraph, KBNode, KBEdge } from '../../types';
-import { BUILT_IN_VIEWS, getView, filterGraphToView } from '../../representation/views';
+import type { KBGraph, KBNode, KBEdge } from '../../../types';
+import { BUILT_IN_VIEWS, getView, filterGraphToView } from '../../views';
 import {
   REPORTS_TO_RELATION,
   isPersonNode,
