+++ src/representation/graph-canvas/nodeRenderer.ts
